+++ src/components/Problems/CodeProblem/PeerHelp/HelpMatch.tsx
@@ -4,9 +4,16 @@ import update from 'immutability-helper';
 import { IHelpSession, ICodeTest } from '../../../../reducers/aggregateData';
 import { ICodeSolutionState } from '../../../../reducers/intermediateUserState';
 import SessionList from './SessionList';
+import getChannelName from '../../../../utils/channelName';
+import { analytics } from '../../../../utils/Firebase';
+import uuid from '../../../../utils/uuid';
+import { addHelpSession, changeHelperLists } from '../../../../actions/sharedb_actions';
+import { updateCurrentActiveHelpSession } from '../../../../actions/user_actions';
 
 
-const HelpMatch = ({ redirectCallback, problem, username, helpSessionObjects, currentTest, currentResult }) => {
+
+
+const HelpMatch = ({ dispatch, redirectCallback, problem, username, myemail, helpSessionObjects, myHelpSession, currentTest, currentResult, userSolution, myuid }) => {
     const [isDisplay, setIsDisplay] = React.useState(true);
 
     let errorTags: string[] = [];
@@ -40,6 +47,39 @@ const HelpMatch = ({ redirectCallback, problem, username, helpSessionObjects, cu
     const toggleDisplay = () => {
         setIsDisplay(!isDisplay)
     }
+    const doRequestHelp = () => {
+        const helpID = uuid();
+        let title = "";
+        let errorTags: string[] = [];
+        let testTags: string[] = [];
+        testTags.push(currentTest.id);
+        if (currentResult && currentResult.passed === "failed") {
+            if (currentResult.errors.length > 0) {
+                const errorMessages = currentResult.errors[0].split('\n');
+                const errorMessage = errorMessages[errorMessages.length - 1];
+                const errorType = errorMessage.split(':')[0];
+                errorTags.push(errorType);
+                if (errorType === "AssertionError") title = "Failed the assertion in **" + currentTest.name + "**";
+                else title = "`" + errorType + "` in my code"
+            }
+        }
+        else {
+            title = "Help me with **" + currentTest.name + "**";
+        }
+        const newCode = currentTest.before + "\n" + userSolution.code + "\n" + currentTest.after;
+        analytics.logEvent("add_help_session", { problemID: problem.id, channel: getChannelName(), user: myemail, helpID });
+
+        dispatch(addHelpSession(problem.id, username, newCode, helpID, errorTags, testTags, title)).then(() => {
+            dispatch(updateCurrentActiveHelpSession(problem.id, helpID));
+            dispatch(changeHelperLists(problem.id, helpID, myuid))
+        });
+        redirectCallback();
+    }
+
+    const doSwitchHelp = () => {
+        dispatch(updateCurrentActiveHelpSession(problem.id, myHelpSession.id))
+        redirectCallback();
+    }
 
     // if the user finished the problem, then directed him to the latest help request
     if ((currentResult && currentResult.passed === "passed") || matchedSessions.length === 0) {
@@ -49,17 +89,26 @@ const HelpMatch = ({ redirectCallback, problem, username, helpSessionObjects, cu
     }
 
     return <div className="match-container">
+        {myHelpSession === null &&
+            <button type="button" className="btn btn-outline-primary" onClick={doRequestHelp}><i className="fas fa-comment"></i> Start a Help Session</button>
+        }
+        {myHelpSession !== null &&
+            <button type="button" className="btn btn-outline-secondary" onClick={doSwitchHelp}>Check My Help Session</button>
+        }
         {(currentResult && currentResult.passed !== "pending") &&
             <>
-                <button type="button" className="btn btn-outline-primary" onClick={toggleDisplay}>{isDisplay ? "Hide Suggestions" : "Display Suggestions"}</button>
+                <div className="custom-control custom-switch related-button">
+                    <input type="checkbox" className="custom-control-input" id={"related-help-button-" + problem.id} onClick={toggleDisplay} defaultChecked={isDisplay} />
+                    <label className="custom-control-label" htmlFor={"related-help-button-" + problem.id}>Related Help Sessions</label>
+                </div>
                 {isDisplay &&
                     <div className="alert alert-light alert-dismissible fade show" role="alert">
                         {currentResult.passed === "passed"
                             ? <h4 className="alert-heading">Could you help them?</h4>
-                            : <h4 className="alert-heading">You may find these discussions useful.</h4>
+                            : <h4 className="alert-heading">You may find these sessions useful.</h4>
                         }
                         {matchedSessions.length === 0 &&
-                            <div>No matched sessions right now. </div>
+                            <div>No related help sessions right now. </div>
                         }
                         {matchedSessions.map((session, i) => <SessionList key={i} session={session} problem={problem} clickCallback={clickCallback} />)}
                     </div>
@@ -70,13 +119,14 @@ const HelpMatch = ({ redirectCallback, problem, username, helpSessionObjects, cu
 }
 
 function mapStateToProps(state, ownProps) {
-    const { intermediateUserState, shareDBDocs, users } = state;
+    const { intermediateUserState, shareDBDocs, solutions, users } = state;
     const { problem } = ownProps;
     const aggregateData = shareDBDocs.i.aggregateData
     const intermediateCodeState: ICodeSolutionState = intermediateUserState.intermediateSolutionState[ownProps.problem.id];
     const { problemDetails } = problem;
     const instructorTests = problemDetails.tests;
     const myuid = users.myuid as string;
+    const myemail = users.allUsers[myuid].email;
     const username = users.allUsers[myuid].username;
 
     const { currentActiveTest, testResults } = intermediateCodeState;
@@ -84,6 +134,8 @@ function mapStateToProps(state, ownProps) {
     if (aggregateData) {
         tests = aggregateData.userData[problem.id].tests;
     }
+    const userSolution = solutions.allSolutions[problem.id][myuid];
+
 
     const instructorTestObjects: ICodeTest[] = Object.values(instructorTests);
     const allTests = Object.assign(JSON.parse(JSON.stringify(tests)), instructorTests);
@@ -97,6 +149,9 @@ function mapStateToProps(state, ownProps) {
     }
     const helpSessionObjects: IHelpSession[] = Object.values(helpSessions);
 
-    return update(ownProps, { $merge: { problem, helpSessionObjects, helpSessions, currentTest, currentResult, username } });
+    let myHelpS = helpSessionObjects.filter(s => s.tutee === username && s.status);
+    const myHelpSession = myHelpS.length > 0 ? myHelpS[0] : null;
+
+    return update(ownProps, { $merge: { helpSessionObjects, currentTest, currentResult, username, myHelpSession, userSolution, myuid, myemail } });
 }
 export default connect(mapStateToProps)(HelpMatch);
